+++ lib/postgres/repository/order.go
@@ -58,3 +58,11 @@ func (r *Repository) DeleteOrder(shopId int32, orderId int32, userId string) err
 	}
 	return nil
 }
+
+func (r *Repository) GetOrderWithShopItemsById(id int32, shopId int32) (*models.Order, error) {
+	var order models.Order
+	if err := r.pg.Where("id = ? AND shop_id = ?", id, shopId).Preload("Items.ShopItem").Find(&order).Error; err != nil {
+		return nil, err
+	}
+	return &order, nil
+}
